feat: Add all profile metadata to the chrometrace output
