+++ deployments/docker-compose.yml
@@ -1,5 +1,4 @@
 version: "3.9"
-name: vroom
 services:
   gcs:
     image: "fsouza/fake-gcs-server:latest"
